+++ Changelog.md
@@ -1,3 +1,7 @@
+# Changelog
+
+This log tracks only major changes, as well as changes that need characters to be updated.
+
 ## 1.0.0
 
 - First published PECS in a way that presupposes that other people could want to use it - with a readme and a version number and all, and without development files included.
@@ -11,3 +15,7 @@
 
 - Monks created before version 1.0.2 will lose their skill increases from Path to Perfection, and gain matching feature choices instead.
 - A Foundry VTT integration allows you to send your dice rolls to a FVTT session. This requires the External Dice Roll Connector module to be enabled in FVTT.
+
+## 1.0.3
+
+- Rogues created before version 1.0.3 will have their "Racket" choice renamed to "Rogue's Racket" to match the correct class choice name.